chore: use createThrottle util
