+++ src/lib/market-service/portals/portals.ts
@@ -22,6 +22,7 @@ import { assertUnreachable, getTimeFrameBounds, isToken } from 'lib/utils'
 import generatedAssetData from '../../asset-service/service/generatedAssetData.json'
 import type { MarketService } from '../api'
 import { DEFAULT_CACHE_TTL_MS } from '../config'
+import { createThrottle } from '../utils'
 import { isValidDate } from '../utils/isValidDate'
 import { CHAIN_ID_TO_PORTALS_NETWORK } from './constants'
 import type { GetTokensResponse, HistoryResponse } from './types'
@@ -33,53 +34,7 @@ const calculatePercentChange = (openPrice: string, closePrice: string): number =
   return close.minus(open).dividedBy(open).times(100).toNumber()
 }
 
-const { throttle, clear } = (({
-  capacity,
-  costPerReq,
-  drainPerInterval,
-  intervalMs,
-}: {
-  capacity: number
-  costPerReq: number
-  drainPerInterval: number
-  intervalMs: number
-}) => {
-  let currentLevel = 0
-  let pendingResolves: ((value?: unknown) => void)[] = []
-
-  const drain = () => {
-    const drainAmount = Math.min(currentLevel, drainPerInterval)
-    currentLevel -= drainAmount
-
-    // Resolve pending promises if there's enough capacity
-    while (pendingResolves.length > 0 && currentLevel + costPerReq <= capacity) {
-      const resolve = pendingResolves.shift()
-      if (resolve) {
-        currentLevel += costPerReq
-        resolve()
-      }
-    }
-  }
-
-  // Start the interval to drain the capacity
-  const intervalId = setInterval(drain, intervalMs)
-
-  const throttle = async () => {
-    if (currentLevel + costPerReq <= capacity) {
-      // If adding another request doesn't exceed capacity, proceed immediately
-      currentLevel += costPerReq
-    } else {
-      // Otherwise, wait until there's enough capacity
-      await new Promise(resolve => {
-        pendingResolves.push(resolve)
-      })
-    }
-  }
-
-  const clear = () => clearInterval(intervalId)
-
-  return { throttle, clear }
-})({
+const { throttle, clear } = createThrottle({
   capacity: 500, // 500 rpm as per https://github.com/shapeshift/web/pull/7401#discussion_r1687499650
   costPerReq: 1,
   drainPerInterval: 125, // Replenish 125 requests every 15 seconds
