+++ src/lib/market-service/zerion/zerion.ts
@@ -12,20 +12,20 @@ import type { MarketService } from '../api'
 export class ZerionMarketService implements MarketService {
   baseUrl = 'https://zerion.shapeshift.com/'
 
-  async findAll(args?: FindAllMarketArgs): Promise<MarketCapResult> {
+  async findAll(_args?: FindAllMarketArgs): Promise<MarketCapResult> {
     // TODO
     return await Promise.resolve({})
   }
 
   async findPriceHistoryByAssetId({
-    assetId,
-    timeframe,
+    assetId: _assetId,
+    timeframe: _timeframe,
   }: PriceHistoryArgs): Promise<HistoryData[]> {
     // TODO
     return await Promise.resolve([])
   }
 
-  async findByAssetId({ assetId }: MarketDataArgs): Promise<MarketData | null> {
+  async findByAssetId({ assetId: _assetId }: MarketDataArgs): Promise<MarketData | null> {
     // TODO
     return await Promise.resolve(null)
   }
